changes to support a 4th axis 'E' and pinout rearrange

+++ grbl/settings.h
@@ -1,10 +1,10 @@
 /*
-  settings.h - eeprom configuration handling 
+  settings.h - eeprom configuration handling
   Part of Grbl
 
   Copyright (c) 2011-2015 Sungeun K. Jeon
   Copyright (c) 2009-2011 Simen Svale Skogsrud
-  
+
   Grbl is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
@@ -24,27 +24,26 @@
 
 #include "grbl.h"
 
-
 // Version of the EEPROM data. Will be used to migrate existing data from older versions of Grbl
 // when firmware is upgraded. Always stored in byte 0 of eeprom
-#define SETTINGS_VERSION 9  // NOTE: Check settings_reset() when moving to next version.
+#define SETTINGS_VERSION 9 // NOTE: Check settings_reset() when moving to next version.
 
 // Define bit flag masks for the boolean settings in settings.flag.
-#define BITFLAG_REPORT_INCHES      bit(0)
+#define BITFLAG_REPORT_INCHES bit(0)
 // #define BITFLAG_AUTO_START         bit(1) // Obsolete. Don't alter to keep back compatibility.
-#define BITFLAG_INVERT_ST_ENABLE   bit(2)
-#define BITFLAG_HARD_LIMIT_ENABLE  bit(3)
-#define BITFLAG_HOMING_ENABLE      bit(4)
-#define BITFLAG_SOFT_LIMIT_ENABLE  bit(5)
-#define BITFLAG_INVERT_LIMIT_PINS  bit(6)
-#define BITFLAG_INVERT_PROBE_PIN   bit(7)
+#define BITFLAG_INVERT_ST_ENABLE bit(2)
+#define BITFLAG_HARD_LIMIT_ENABLE bit(3)
+#define BITFLAG_HOMING_ENABLE bit(4)
+#define BITFLAG_SOFT_LIMIT_ENABLE bit(5)
+#define BITFLAG_INVERT_LIMIT_PINS bit(6)
+#define BITFLAG_INVERT_PROBE_PIN bit(7)
 
 // Define status reporting boolean enable bit flags in settings.status_report_mask
-#define BITFLAG_RT_STATUS_MACHINE_POSITION  bit(0)
-#define BITFLAG_RT_STATUS_WORK_POSITION     bit(1)
-#define BITFLAG_RT_STATUS_PLANNER_BUFFER    bit(2)
-#define BITFLAG_RT_STATUS_SERIAL_RX         bit(3)
-#define BITFLAG_RT_STATUS_LIMIT_PINS        bit(4)
+#define BITFLAG_RT_STATUS_MACHINE_POSITION bit(0)
+#define BITFLAG_RT_STATUS_WORK_POSITION bit(1)
+#define BITFLAG_RT_STATUS_PLANNER_BUFFER bit(2)
+#define BITFLAG_RT_STATUS_SERIAL_RX bit(3)
+#define BITFLAG_RT_STATUS_LIMIT_PINS bit(4)
 
 // Define settings restore bitflags.
 #define SETTINGS_RESTORE_ALL 0xFF // All bitflags
@@ -55,28 +54,29 @@
 
 // Define EEPROM memory address location values for Grbl settings and parameters
 // NOTE: The Atmega328p has 1KB EEPROM. The upper half is reserved for parameters and
-// the startup script. The lower half contains the global settings and space for future 
+// the startup script. The lower half contains the global settings and space for future
 // developments.
-#define EEPROM_ADDR_GLOBAL         1U
-#define EEPROM_ADDR_PARAMETERS     512U
-#define EEPROM_ADDR_STARTUP_BLOCK  768U
-#define EEPROM_ADDR_BUILD_INFO     942U
+#define EEPROM_ADDR_GLOBAL 1U
+#define EEPROM_ADDR_PARAMETERS 512U
+#define EEPROM_ADDR_STARTUP_BLOCK 768U
+#define EEPROM_ADDR_BUILD_INFO 942U
 
 // Define EEPROM address indexing for coordinate parameters
-#define N_COORDINATE_SYSTEM 6  // Number of supported work coordinate systems (from index 1)
-#define SETTING_INDEX_NCOORD N_COORDINATE_SYSTEM+1 // Total number of system stored (from index 0)
+#define N_COORDINATE_SYSTEM 6                        // Number of supported work coordinate systems (from index 1)
+#define SETTING_INDEX_NCOORD N_COORDINATE_SYSTEM + 1 // Total number of system stored (from index 0)
 // NOTE: Work coordinate indices are (0=G54, 1=G55, ... , 6=G59)
-#define SETTING_INDEX_G28    N_COORDINATE_SYSTEM    // Home position 1
-#define SETTING_INDEX_G30    N_COORDINATE_SYSTEM+1  // Home position 2
+#define SETTING_INDEX_G28 N_COORDINATE_SYSTEM     // Home position 1
+#define SETTING_INDEX_G30 N_COORDINATE_SYSTEM + 1 // Home position 2
 // #define SETTING_INDEX_G92    N_COORDINATE_SYSTEM+2  // Coordinate offset (G92.2,G92.3 not supported)
 
 // Define Grbl axis settings numbering scheme. Starts at START_VAL, every INCREMENT, over N_SETTINGS.
-#define AXIS_N_SETTINGS          4
-#define AXIS_SETTINGS_START_VAL  100 // NOTE: Reserving settings values >= 100 for axis settings. Up to 255.
-#define AXIS_SETTINGS_INCREMENT  10  // Must be greater than the number of axis settings
+#define AXIS_N_SETTINGS 4
+#define AXIS_SETTINGS_START_VAL 100 // NOTE: Reserving settings values >= 100 for axis settings. Up to 255.
+#define AXIS_SETTINGS_INCREMENT 10  // Must be greater than the number of axis settings
 
 // Global persistent settings (Stored from byte EEPROM_ADDR_GLOBAL onwards)
-typedef struct {
+typedef struct
+{
   // Axis settings
   float steps_per_mm[N_AXIS];
   float max_rate[N_AXIS];
@@ -88,11 +88,11 @@ typedef struct {
   uint8_t step_invert_mask;
   uint8_t dir_invert_mask;
   uint8_t stepper_idle_lock_time; // If max value 255, steppers do not disable.
-  uint8_t status_report_mask; // Mask to indicate desired report data.
+  uint8_t status_report_mask;     // Mask to indicate desired report data.
   float junction_deviation;
   float arc_tolerance;
-  
-  uint8_t flags;  // Contains default boolean settings
+
+  uint8_t flags; // Contains default boolean settings
 
   uint8_t homing_dir_mask;
   float homing_feed_rate;
@@ -130,13 +130,12 @@ void settings_write_coord_data(uint8_t coord_select, float *coord_data);
 uint8_t settings_read_coord_data(uint8_t coord_select, float *coord_data);
 
 // Returns the step pin mask according to Grbl's internal axis numbering
-uint8_t get_step_pin_mask(uint8_t i);
+uint16_t get_step_pin_mask(uint8_t i);
 
 // Returns the direction pin mask according to Grbl's internal axis numbering
-uint8_t get_direction_pin_mask(uint8_t i);
+uint16_t get_direction_pin_mask(uint8_t i);
 
 // Returns the limit pin mask according to Grbl's internal axis numbering
 uint8_t get_limit_pin_mask(uint8_t i);
-
 
 #endif
